% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biodiversity_intactness_index.R,
%   R/get_biodiversity_intactness_index.R
\docType{data}
\name{biodiversity_intactness_index}
\alias{biodiversity_intactness_index}
\alias{get_biodiversity_intactness_index}
\title{Calculate Biodiversity Intactness Index}
\format{
A function returning a tibble with the aggregated biodiversity
intactness index.
}
\source{
\url{https://data.nhm.ac.uk/dataset/global-map-of-the-biodiversity-intactness-index-from-newbold-et-al-2016-science}
}
\usage{
get_biodiversity_intactness_index(path = NULL)
}
\arguments{
\item{path}{A character vector to the biodiversity intactness index tif file.
Note, that the file has to be downloaded manually.}
}
\value{
A function that returns a character vector of file paths.
}
\description{
This function calculates the mean biodiversity intactness index for a region.

The variable is the modeled average abundance of originally-present species,
relative to their abundance in an intact ecosystem. Please refer to
Newbold et al. (2016) for all details, and please cite it when using these
data.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{biodiversity_intactness_index}
}

To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.
Please find the available data under the source link given below.
}
\references{
Tim Newbold; Lawrence Hudson; Andy Arnell; Sara Contu et al. (2016). Global map of the Biodiversity Intactness Index, from Newbold et al. (2016) Science [Data set]. Natural History Museum. \doi{https://doi.org/10.5519/0009936}
}
\keyword{indicator}
\keyword{resource}
