% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn.R
\docType{data}
\name{birdlife}
\alias{birdlife}
\title{BirdLife International Species Ranges}
\format{
A global vector data set indicating species ranges
}
\source{
\url{http://datazone.birdlife.org/species/requestdis}
}
\description{
This resource is part of the spatial data set of bird ranges
released by BirdLife International. This resource is free to use under a
non-commercial licence. For commercial uses, a request has to be sent to
BirdLife International.
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.

Note, that we are simplifying the species ranges polygons for better computational
efficiency.
}
\references{
BirdLife International and Handbook of the Birds of the World (2022).
Bird species distribution maps of the world. Version 2022.2. Available at
http://datazone.birdlife.org/species/requestdis
}
\keyword{resource}
