% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw_time_series.R
\docType{data}
\name{gsw_time_series}
\alias{gsw_time_series}
\title{Global Surface Water - Yearly Time Series}
\format{
A global tiled raster resource available for all terrestrial areas, available yearly for 1984 -- 2021.
}
\source{
Raw Data: \url{https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GSWE/YearlyClassification/LATEST/tiles/}
}
\description{
This resource is part of the JRC Global Surface Water dataset. It provides
yearly categorical raster data on terrestrial surface water classes.
}
\details{
The available surface water classes for a given pixel are the following:
\itemize{
\item No Observation: It was not possible to determine whether a pixel was water
(this may be the case for frozen areas or during the polar night in extreme
latitudes).
\item Permanent Water: Water was detected in twelve months per year or in a
combination of permanent and no observation.
\item Seasonal Water: Water and no water was detected.
\item No Water: No Water was detected.
}
}
\references{
\itemize{
\item Global Surface Water Explorer: \url{https://global-surface-water.appspot.com/}
\item Data Users Guide: \url{https://storage.cloud.google.com/global-surface-water/downloads_ancillary/DataUsersGuidev2021.pdf}
\item Research Article: \url{https://www.nature.com/articles/nature20584}
}
}
\keyword{resource}
