% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hfp.R
\docType{data}
\name{humanfootprint}
\alias{humanfootprint}
\title{Terrestrial Human Foootprint}
\format{
A global tiled raster resource available for all land areas.
}
\source{
\url{https://figshare.com/articles/figure/An_annual_global_terrestrial_Human_Footprint_dataset_from_2000_to_2018/16571064}
}
\description{
This resource is part of the publication by Mu et al. (2022) "A global
record of annual terrestrial Human Footprint dataset from 2000 to 2018".
}
\references{
Mu, H., Li, X., Wen, Y. et al. A global record of annual
terrestrial Human Footprint dataset from 2000 to 2018. Sci Data 9, 176 (2022).
https://doi.org/10.1038/s41597-022-01284-8
}
\keyword{resource}
