% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ipbes.R, R/get_ipbes.R
\docType{data}
\name{ipbes_anthrome}
\alias{ipbes_anthrome}
\title{Calculate areal statistics for IBPES Anthromes}
\format{
A tibble with columns for the respective anthrome name, the affected
area, the percentage of the affected area in relation to the total area,
and the total area (all expressed in ha).

A global tiled raster resource available for all land areas.
}
\source{
\url{https://zenodo.org/records/3975694}
}
\description{
This indicator calculates the areal distribution of the two different anthrome
classes within an asset based on the IBPES anthrome dataset. You can select
to only include areal statistics for built-up areas, cultivated areas, or both.

This resource is part of the Global Assessment Report on Biodiversity and
Ecosystem Services and represents a division of the Earth's surface
into several subcategories. The classification differentiates between
biomes and anthromes. Anthromes are differentiated between two classes,
which are built-up (urbanized) areas and cultivated areas.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{ipbes_anthrome}
}

The following argument can be set:
\describe{
\item{anthrome}{One of "urban_areas", "cultivated_areas", "both". Determines
for which data layer the areal statistics are calculated.}
}
}
\references{
IPBES (2019): Summary for policymakers of the global assessment
report on biodiversity and ecosystem services of the Intergovernmental
Science-Policy Platform on Biodiversity and Ecosystem Services. S. Díaz, J.
Settele, E. S. Brondízio, H. T. Ngo, M. Guèze, J. Agard, A. Arneth, P.
Balvanera, K. A. Brauman, S. H. M. Butchart, K. M. A. Chan, L. A. Garibaldi,
K. Ichii, J. Liu, S. M. Subramanian, G. F. Midgley, P. Miloslavich, Z.
Molnár, D. Obura, A. Pfaff, S. Polasky, A. Purvis, J. Razzaque, B. Reyers, R.
Roy Chowdhury, Y. J. Shin, I. J. Visseren-Hamakers, K. J. Willis, and C. N.
Zayas (eds.). IPBES secretariat, Bonn, Germany. 56 pages.
https://doi.org/10.5281/zenodo.3553579
}
\keyword{indicator}
\keyword{resource}
