% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ipbes.R
\name{ipbes_anthrome_stats}
\alias{ipbes_anthrome_stats}
\alias{calc_ipbes_anthrome}
\title{Calculate areal statistics for IBPES Anthromes}
\usage{
calc_ipbes_anthrome(anthrome = c("urban_areas", "cultivated_areas", "both"))
}
\arguments{
\item{anthrome}{One of "urban_areas", "cultivated_areas", "both". Determines
for which data layer the areal statistics are calculated.}
}
\value{
A function that returns a tibble with columns for the respective
anthrome name, the affected area, the percentage of the affected area in
relation to the total area, and the total area (all expressed in ha).
}
\description{
This indicator calculates the areal distribution of the two different anthrome
classes within an asset based on the IBPES anthrome dataset. You can select
to only include areal statistics for built-up areas, cultivated areas, or both.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{ipbes}
}
}
\keyword{indicator}
