% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon.R
\docType{data}
\name{irr_carbon}
\alias{irr_carbon}
\title{Irrecoverable Carbon Layer}
\format{
A global tiled raster resource available for all land areas.
}
\source{
\url{https://zenodo.org/records/4091029}
}
\description{
This resource is part of the publication by Noon et al. (2022) "Mapping the
irrecoverable carbon in Earth’s ecosystems". This publication differentiates
between 3 different kinds of carbon with varying degrees of manageability
by humans. All three layers are available for above and below ground
carbon, as well as a layer combining the two.
}
\details{
Irrecoverable carbon is defined as the amount of carbon, that, if lost today,
cannot be recovered until mid 21st century (so within 30 years, considering
the publication date).
}
\references{
Noon, M.L., Goldstein, A., Ledezma, J.C. et al. Mapping the
irrecoverable carbon in Earth’s ecosystems. Nat Sustain 5, 37–46 (2022).
https://doi.org/10.1038/s41893-021-00803-6
}
\keyword{resource}
