% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn.R
\docType{data}
\name{iucn_mammals}
\alias{iucn_mammals}
\title{IUCN Red List of Threatened Species - Mammals}
\format{
A global vector data set indicating species ranges
}
\source{
\url{https://www.iucnredlist.org/resources/spatial-data-download}
}
\description{
This resource is part of the spatial data set Red List of Threatened Species
released by IUCN. The data are made available including taxonomic information,
distribution status, IUCN Red List Category, sources and other details. This
resource is free to use under a non-commercial licence. For commercial uses,
a request has to be sent to Integrated Biodiversity Assessment Tool (IBAT).
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.

Note, that we are simplifying the species ranges polygons for better computational
efficiency.
}
\references{
IUCN (2024). The IUCN Red List of Threatened Species. https://www.iucnredlist.org.
https://doi.org/10.1038/s41597-022-01284-8
}
\keyword{resource}
