% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_key_biodiversity_areas.R,
%   R/get_key_biodiversity_areas.R
\docType{data}
\name{key_biodiversity_areas}
\alias{key_biodiversity_areas}
\alias{get_key_biodiversity_areas}
\title{Calculate Key Biodiversity Areas}
\format{
A function returning a tibble with the key biodiversity area overlap
}
\source{
\url{https://www.keybiodiversityareas.org/kba-data}
}
\usage{
get_key_biodiversity_areas(path = NULL)
}
\arguments{
\item{path}{A character vector to the key biodiversity areas gpkg file.
Note, that the file has to be downloaded manually.}
}
\value{
A function that returns a character vector of file paths.
}
\description{
This function calculates the total area of key biodiversity areas for a given
input polygon.

The variable contains outlines of key biodiversity areas. These areas
represent sites with specific importance for nature conservation efforts.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{key_biodiversity_areas}
}

To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.
Please find the available data under the source link given below.
}
\references{
BirdLife International (2024). The World Database of Key Biodiversity Areas. Developed by the KBA Partnership: BirdLife International, International Union for the Conservation of Nature, Amphibian Survival Alliance, Conservation International, Critical Ecosystem Partnership Fund, Global Environment Facility, Re:wild, NatureServe, Rainforest Trust, Royal Society for the Protection of Birds, Wildlife Conservation Society and World Wildlife Fund. Available at www.keybiodiversityareas.org.
}
\keyword{indicator}
\keyword{resource}
