% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn.R
\name{species_ranges}
\alias{species_ranges}
\alias{get_iucn_mammals}
\alias{get_iucn_reptiles}
\alias{get_iucn_amphibians}
\alias{get_birdlife}
\title{IUCN Red List of Threatened Species - Mammals}
\source{
\url{https://www.iucnredlist.org/resources/spatial-data-download}
}
\usage{
get_iucn_mammals(path = NULL)

get_iucn_reptiles(path = NULL)

get_iucn_amphibians(path = NULL)

get_birdlife(path = NULL)
}
\arguments{
\item{path}{A character vector to the respective species range file that
has to be downloaded manually.}
}
\value{
A function that returns a character vector of file paths.
}
\description{
This resource is part of the spatial data set Red List of Threatened Species
released by IUCN. The data are made available including taxonomic information,
distribution status, IUCN Red List Category, sources and other details. This
resource is free to use under a non-commercial licence. For commercial uses,
a request has to be sent to Integrated Biodiversity Assessment Tool (IBAT).
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.

Note, that we are simplifying the species ranges polygons for better computational
efficiency.
}
\references{
IUCN (2024). The IUCN Red List of Threatened Species. https://www.iucnredlist.org.
\doi{https://doi.org/10.1038/s41597-022-01284-8 }

BirdLife International and Handbook of the Birds of the World (2022).
Bird species distribution maps of the world. Version 2022.2. Available at
http://datazone.birdlife.org/species/requestdis
}
\keyword{resource}
