% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iucn.R
\name{species_richness}
\alias{species_richness}
\alias{calc_species_richness}
\title{Species richness based on IUCN raster data}
\usage{
calc_species_richness(
  engine = "extract",
  stats = "mean",
  variable = "species_richness"
)
}
\arguments{
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}

\item{variable}{A character used as the variable column output. Defaults to
\code{species_richness} but can be set to reflect the manually specified raster
file.}
}
\value{
A function that returns a tibble with statistics on the species
richness.
}
\description{
Species richness counts the number of potential species intersecting with a
polygon grouped by the IUCN threat categorization. Note, that
this indicator function requires the manual download of the respective
raster files.
}
\details{
The specific meaning of the species richness indicator depends on the
supplied raster file.

The required resources for this indicator are:
\itemize{
\item \link{iucn}
}
}
\keyword{indicator}
