% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iucn.R
\name{sr_amphibians}
\alias{sr_amphibians}
\alias{calc_sr_amphibians}
\title{Species richness}
\usage{
calc_sr_amphibians()
}
\value{
A function that returns a tibble with columns for the code and the
full-text threat category as well as the count of the number of respective species.
}
\description{
Species richness counts the number of potential species intersecting with a
polygon grouped by the IUCN threat categorization. For each animal group,
there is an indicator function which requires the availablity of the
respctive resource.
}
\details{
The required resources for this indicator are:
\itemize{
\item see the respective \link{species_ranges}
}
}
