% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iucn.R
\docType{data}
\name{sr_birds}
\alias{sr_birds}
\title{Species richness - Birds}
\format{
A tibble with a column for the number of bird species which ranges
intersect with a given polygon.
}
\description{
Species richness counts the number of species intersecting with a polygon.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{birdlife}
}
}
\keyword{indicator}
