% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon.R
\name{vul_carbon}
\alias{vul_carbon}
\alias{get_vul_carbon}
\title{Vulnerable Carbon Layer}
\source{
\url{https://zenodo.org/records/4091029}
}
\usage{
get_vul_carbon()
}
\value{
A function that returns a character vector of file paths.
}
\description{
This resource is part of the publication by Noon et al. (2022) "Mapping the
irrecoverable carbon in Earth’s ecosystems". This publication differentiates
between 3 different kinds of carbon with varying degrees of manageability
by humans. All three layers are available for above and below ground
carbon, as well as a layer combining the two.
}
\details{
Vulnerable carbon is defined as the amount of carbon that would be lost in a
hypothetical but typical conversion event (without including information
of the probability of such an event to be actually occurring).
}
\references{
Noon, M.L., Goldstein, A., Ledezma, J.C. et al. Mapping the
irrecoverable carbon in Earth’s ecosystems. Nat Sustain 5, 37–46 (2022).
https://doi.org/10.1038/s41893-021-00803-6
}
\keyword{resource}
