% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_carbon.R
\name{vul_carbon_stats}
\alias{vul_carbon_stats}
\alias{calc_vul_carbon}
\title{Calculate vulnerable carbon statistics}
\usage{
calc_vul_carbon(
  type = c("total", "soil", "biomass", "all"),
  engine = "extract",
  stats = "mean"
)
}
\arguments{
\item{type}{One of "total", "soil", "biomass", "all". Determines
for which data layer the statistics are calculated.}

\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
}
\value{
A function that returns a tibble with a column for each statistic
and rows for every year and type of carbon.
}
\description{
Vulnerable carbon is the amount of carbon that would be released in a typical
land conversion activity.
It can be calculated for above- and below-ground carbon, the total amount
of carbon, or for all layers.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{vul_carbon}
}
}
\keyword{indicator}
