% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hfp.R
\docType{data}
\name{humanfootprint_stats}
\alias{humanfootprint_stats}
\title{Calculate human footprint statistics}
\format{
A tibble with a column for each statistic and a row for every
requested year.
}
\description{
Human footprint data measures the pressure imposed on the natural environment
by different dimensions of human actions. The theoretical maximum value,
representing the highest level of human pressure, is 50. This routine allows
to extract zonal statistics of the human footprint data.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{humanfootprint}
}
}
\keyword{indicator}
