% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipbes.R
\docType{data}
\name{ipbes_anthrome}
\alias{ipbes_anthrome}
\title{Terrestrial Anthromes}
\format{
A global tiled raster resource available for all land areas.
}
\source{
\url{https://zenodo.org/records/3975694}
}
\description{
This resource is part of the Global Assessment Report on Biodiversity and
Ecosystem Services.
}
\references{
IPBES (2019): Summary for policymakers of the global assessment
report on biodiversity and ecosystem services of the Intergovernmental
Science-Policy Platform on Biodiversity and Ecosystem Services. S. Díaz, J.
Settele, E. S. Brondízio, H. T. Ngo, M. Guèze, J. Agard, A. Arneth, P.
Balvanera, K. A. Brauman, S. H. M. Butchart, K. M. A. Chan, L. A. Garibaldi,
K. Ichii, J. Liu, S. M. Subramanian, G. F. Midgley, P. Miloslavich, Z.
Molnár, D. Obura, A. Pfaff, S. Polasky, A. Purvis, J. Razzaque, B. Reyers, R.
Roy Chowdhury, Y. J. Shin, I. J. Visseren-Hamakers, K. J. Willis, and C. N.
Zayas (eds.). IPBES secretariat, Bonn, Germany. 56 pages.
https://doi.org/10.5281/zenodo.3553579
}
\keyword{resource}
