% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ipbes.R
\docType{data}
\name{ipbes_anthrome_stats}
\alias{ipbes_anthrome_stats}
\title{Calculate areal statistics for IBPES Anthromes}
\format{
A tibble with columns for the respective anthrome name, the affected
area, the percentage of the affected area in relation to the total area,
and the total area (all expressed in ha).
}
\description{
This indicator calculates the areal distribution of the two different anthrome
classes within an asset based on the IBPES anthrome dataset. You can select
to only include areal statistics for built-up areas, cultivated areas, or both.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{ipbes_anthrome}
}

The following argument can be set:
\describe{
\item{anthrome}{One of "urban_areas", "cultivated_areas", "both". Determines
for which data layer the areal statistics are calculated.}
}
}
\keyword{indicator}
