% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_carbon.R, R/get_carbon.R
\docType{data}
\name{man_carbon}
\alias{man_carbon}
\title{Calculate manageable carbon statistics}
\format{
A tibble with a column for each statistic and rows for every year
and type of carbon.

A global tiled raster resource available for all land areas.
}
\source{
\url{https://zenodo.org/records/4091029}
}
\description{
Manageable carbon is the amount of carbon that, in principle, is manageable
by human activities, e.g. its release to the atmosphere can be prevented.
It can be calculated for above- and below-ground carbon, the total amount
of carbon, or for all layers.

This resource is part of the publication by Noon et al. (2022) "Mapping the
irrecoverable carbon in Earth’s ecosystems". This publication differentiates
between 3 different kinds of carbon with varying degrees of manageability
by humans. All three layers are available for above and below ground
carbon, as well as a layer combining the two.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{man_carbon}
}

The following argument can be set:
\describe{
\item{which_carbon}{One of "total", "soil", "biomass", "all". Determines
for which data layer the statistics are calculated.}
\item{stats_carbon}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}

Manageable carbon is defined as all land areas, expect cyrosols, because
carbon loss is driven by direct land-use conversion which could be halted or
because climate change impacts affecting the area can potentially be directly
mitigated through adaptive management.
}
\references{
Noon, M.L., Goldstein, A., Ledezma, J.C. et al. Mapping the
irrecoverable carbon in Earth’s ecosystems. Nat Sustain 5, 37–46 (2022).
https://doi.org/10.1038/s41893-021-00803-6
}
\keyword{indicator}
\keyword{resource}
