% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_carbon.R
\docType{data}
\name{man_carbon_stats}
\alias{man_carbon_stats}
\title{Calculate manageable carbon statistics}
\format{
A tibble with a column for each statistic and rows for every year
and type of carbon.
}
\description{
Manageable carbon is the amount of carbon that, in principle, is manageable
by human activities, e.g. its release to the atmosphere can be prevented.
It can be calculated for above- and below-ground carbon, the total amount
of carbon, or for all layers.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{man_carbon}
}

The following argument can be set:
\describe{
\item{which_carbon}{One of "total", "soil", "biomass", "all". Determines
for which data layer the statistics are calculated.}
\item{stats_carbon}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\keyword{indicator}
