% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iucn.R
\docType{data}
\name{sr_amphibians}
\alias{sr_amphibians}
\title{Species richness - Amphibians}
\format{
A tibble with columns for the code and full-text category as well
as the count of the number of respective species.
}
\description{
Species richness counts the number of species intersecting with a polygon
grouped by the IUCN threat categorization.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{iucn_amphibians}
}
}
\keyword{indicator}
